Add a frenzy event.
This commit is a pretty good indicator for how to look at events on a
stage level. We tell the Enemy to send an "enemy_killed" event to the
stage whenever it dies, and the stage interprets the event, and triggers
the 'frenzy' event on all enemies once a certain amount of kills have
happened.